战绩得分改为浮点数型

+++ backend/src/main/scala/com/neo/sk/carnie/protocol/EsheepProtocol.scala
@@ -49,7 +49,7 @@ object EsheepProtocol {
                          nickname: String,
                          killing: Int,
                          killed: Int,
-                         score: Int,
+                         score: Float,
                          gameExtent: String,
                          startTime: Long,
                          endTime: Long
